Update Atomist dependencies, remove packs
Update to latest master branch of automation-client, sdm, sdm-core,
and sdm-pack-gcp.  This breaks many packs, so remove them and
duplicate the needed capabilities with container goals.  A side
benefit is that this SDM is one step closer to being a set of skills.

Remove use of goal, goalCreator, and goalConfigurer.